Add ability to choose between files or no files during import
This refactors the update code paths so that they can be
use different `Updater` classes. The specific class used
is chosen in the UI via a select box.

This implements the first two paths: metadata only and full
(update metadata and files)

This also completes the system spec for the full import process
and adds the update path as well.

Since that was working the other importer specific system specs
from tenejo are also added.
