+++ .gitignore
@@ -14,3 +14,4 @@ auto-save-list
 tramp
 .\#*
 coverage/
+*.gem
