+++ app/views/zizia/csv_imports/_form.html.erb
@@ -51,6 +51,14 @@
           </div>
           <div class="col-md-4">
             <div class="well well-lg">
+              <p> <b>Import Behavior</b>. Choose how you would like existing IDs to be handled.</p>
+              <div class="text-center">
+                <%= render "update_actor_stack", form: form %>
+              </div>
+            </div>
+          </div>
+          <div class="col-md-4">
+            <div class="well well-lg">
               <p> <b>Preview your import</b> before starting the process. </p>
               <div class="text-center">
                 <%= render "actions", form: form %>
