+++ app/views/zizia/csv_imports/_update_actor_stack.html.erb
@@ -1,0 +1,5 @@
+<div class="form-group">
+  <%= form.select :update_actor_stack, options_for_select([['Update metadata for any existing IDs', 'HyraxMetadataOnly'],
+                                                           ['Only create works for new IDs', 'HyraxOnlyNew' ],
+                                                           ['Update metadata and files or create new works as required', 'HyraxDefault']]), {}, class: "form-control", id: "update_actor_stack_id"  %>
+</div>
