+++ db/migrate/201901162141281_create_zizia_csv_imports.rb
@@ -3,7 +3,6 @@ class CreateZiziaCsvImports < ActiveRecord::Migration[5.1]
   def change
     create_table :zizia_csv_imports do |t|
       t.references :user, foreign_key: true
-
       t.timestamps
     end
   end
