+++ db/migrate/201901241536542_add_update_actor_stack_to_zizia_csv_imports.rb
@@ -1,0 +1,5 @@
+class AddUpdateActorStackToZiziaCsvImports < ActiveRecord::Migration[5.1]
+  def change
+    add_column :zizia_csv_imports, :update_actor_stack, :string
+  end
+end
