+++ db/migrate/20190921153802_add_depositor_id_to_csv_import_detail.rb
@@ -1,0 +1,5 @@
+class AddDepositorIdToCsvImportDetail < ActiveRecord::Migration[5.1]
+  def change
+    add_column :zizia_csv_import_details, :depositor_id, :string
+  end
+end
