+++ db/migrate/20190921153851_add_collection_id_to_csv_import_detail.rb
@@ -1,0 +1,5 @@
+class AddCollectionIdToCsvImportDetail < ActiveRecord::Migration[5.1]
+  def change
+    add_column :zizia_csv_import_details, :collection_id, :string
+  end
+end
