+++ db/migrate/20190921153933_add_batch_id_to_csv_import_detail.rb
@@ -1,0 +1,5 @@
+class AddBatchIdToCsvImportDetail < ActiveRecord::Migration[5.1]
+  def change
+    add_column :zizia_csv_import_details, :batch_id, :string
+  end
+end
