+++ db/migrate/20190921154009_add_success_count_to_csv_import_detail.rb
@@ -1,0 +1,5 @@
+class AddSuccessCountToCsvImportDetail < ActiveRecord::Migration[5.1]
+  def change
+    add_column :zizia_csv_import_details, :success_count, :integer
+  end
+end
