+++ db/migrate/20190921154038_add_failure_count_to_csv_import_detail.rb
@@ -1,0 +1,5 @@
+class AddFailureCountToCsvImportDetail < ActiveRecord::Migration[5.1]
+  def change
+    add_column :zizia_csv_import_details, :failure_count, :integer
+  end
+end
