+++ db/migrate/20190921154221_add_deduplication_field_to_csv_import_detail.rb
@@ -1,0 +1,5 @@
+class AddDeduplicationFieldToCsvImportDetail < ActiveRecord::Migration[5.1]
+  def change
+    add_column :zizia_csv_import_details, :deduplication_field, :string
+  end
+end
