+++ db/migrate/20190923182946_add_update_actor_stack_to_csv_import_detail.rb
@@ -1,0 +1,5 @@
+class AddUpdateActorStackToCsvImportDetail < ActiveRecord::Migration[5.1]
+  def change
+    add_column :zizia_csv_import_details, :update_actor_stack, :string
+  end
+end
