+++ spec/dummy/app/actors/hyrax/actors/work_actor.rb
@@ -1,0 +1,8 @@
+# Generated via
+#  `rails generate hyrax:work Work`
+module Hyrax
+  module Actors
+    class WorkActor < Hyrax::Actors::BaseActor
+    end
+  end
+end
