--- spec/dummy/app/actors/work_actor.rb
@@ -1,8 +1,0 @@
-# Generated via
-#  `rails generate hyrax:work Work`
-module Hyrax
-  module Actors
-    class WorkActor < Hyrax::Actors::BaseActor
-    end
-  end
-end
