+++ spec/dummy/config/environments/test.rb
@@ -40,4 +40,5 @@ Rails.application.configure do
 
   # Raises error for missing translations
   # config.action_view.raise_on_missing_translations = true
+  config.active_job.queue_adapter = :test
 end
