+++ spec/dummy/spec/fixtures/csv_import/good/all_fields_metadata_update.csv
@@ -1,0 +1,2 @@
+identifier,license,visibility,location,keyword,rights statement,creator,title,files
+abc/123,https://creativecommons.org/licenses/by/4.0/,PUBlic,http://www.geonames.org/5667009/montana.html|~|http://www.geonames.org/6252001/united-states.html,Clothing stores $z California $z Los Angeles|~|Interior design $z California $z Los Angeles,http://rightsstatements.org/vocab/InC/1.0/,"Connell, Will, $d 1898-1961","Exterior view of The Bachelors haberdashery designed by Julius Ralph Davidson, Los Angeles, circa 1929",dog.jpg
