+++ spec/fixtures/csv_import/csv_files_with_problems/missing_headers.csv
@@ -1,0 +1,2 @@
+representative media,thumbnail,rendering,depositor,date_uploaded,date_modified,label,relative_path,import url,resource type,contributor,abstract or summary,license,publisher,date created,subject,language,identifier,location,related url,bibliographic_citation,source
+,,,,,,Work 1 label,,,,,,,,,,,,,,,,
